Enforce comma-separated terms
This change makes the design consistent.

Firstly, variadic arguments provided to desugaring macros must be comma-separated as well as ordinary arguments from now and then.

Secondly, a single term can be provided to some macro without a fear that a comma after it (the previous implementation hack) may affect something.

Thirdly, arguments to METALANG99_call are now separated by commas, which is a more natural C-style, as well as better code formatting with clang-format.